Turn add section UX into modal flow
At long last we have a modal-driven add section process @HarveyKandola !
