+++ gui/app/components/section/code/type-editor.js
@@ -10,13 +10,12 @@
 // https://documize.com
 
 import { computed } from '@ember/object';
-import Component from '@ember/component';
 import TooltipMixin from '../../../mixins/tooltip';
+import Component from '@ember/component';
 
 export default Component.extend(TooltipMixin, {
     isDirty: false,
     pageBody: "",
-
     codeSyntax: null,
 	codeEditor: null,
 	editorId: computed('page', function () {
