+++ gui/app/styles/app.scss
@@ -25,11 +25,3 @@
 @import "news.scss";
 @import "section/all.scss";
 @import "enterprise/all.scss";
-
-// Bootstrap override that removes gutter space on smaller screens
-// @media (max-width: 1200px) {
-//     .container {
-//         width: 100%;
-//         max-width: none;
-//     }
-// }
