+++ gui/app/templates/components/document/add-section.hbs
@@ -1,0 +1,64 @@
+<div id="add-section-modal" class="modal" tabindex="-1" role="dialog">
+	<div class="modal-dialog modal-80" role="document">
+		<div class="modal-content">
+			<div class="modal-header">Insert Section</div>
+			<div class="modal-body">
+
+				<div id="new-section-wizard" class="new-section-wizard">
+					<div class="container box">
+
+						<div class="row">
+							<div class="col-12">
+								<div class="form-group">
+									{{focus-input type="text" id="new-section-name" value=newSectionName
+										class=(if newSectionNameMissing 'mousetrap form-control form-control-lg	is-invalid' 'mousetrap form-control form-control-lg')
+										placeholder="Enter section name" autocomplete="off"}}
+								</div>
+							</div>
+						</div>
+
+						<div class="row">
+							<div class="col-12 col-md-6">
+								<div class="form-group">
+									<div class="new-section-caption">Select Section Type</div>
+									<ul class="preset-list">
+										{{#each sections as |section|}}
+										<li class="item" {{action 'onInsertSection' section}}>
+											<div class="icon">
+												<img class="img" src="/sections/{{section.contentType}}.png" srcset="/sections/{{section.contentType}}@2x.png" />
+											</div>
+											<div class='title'>{{section.title}}</div>
+										</li>
+										{{/each}}
+									</ul>
+								</div>
+							</div>
+							<div class="col-12 col-md-6">
+								{{#if hasBlocks}}
+									<div class="new-section-caption">Select Re-usable Content</div>
+									<ul class="block-list">
+										{{#each blocks as |block|}}
+											<li class="item" title="{{block.firstname}} {{block.lastname}}, {{time-ago block.created}}, used: {{ block.used }}" data-toggle="tooltip" data-placement="top">
+												<div class="details" {{action 'onInsertBlock' block}}>
+													<div class="title text-truncate">{{block.title}}</div>
+													<div class="desc text-truncate">{{block.excerpt}}</div>
+												</div>
+											</li>
+										{{/each}}
+									</ul>
+								{{else}}
+									<div class="template-caption">You have no reusable content &mdash; publish any section as a template</div>
+								{{/if}}
+							</div>
+						</div>
+
+					</div>
+				</div>
+
+			</div>
+			<div class="modal-footer">
+				<button type="button" class="btn btn-secondary" data-dismiss="modal">Cancel</button>
+			</div>
+		</div>
+	</div>
+</div>
